Cleanups and formatting
